# `querylog`: json format version 2