--- a/go/vt/vttablet/tabletserver/tabletenv/logstats.go
+++ b/go/vt/vttablet/tabletserver/tabletenv/logstats.go
@@ -18,6 +18,7 @@ package tabletenv
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"io"
 	"net/url"
@@ -43,13 +44,13 @@ const (
 
 // LogStats records the stats for a single query
 type LogStats struct {
-	Ctx                  context.Context
+	Ctx                  context.Context `json:"-"`
 	Method               string
-	Target               *querypb.Target
+	Target               *querypb.Target `json:"-"`
 	PlanType             string
 	OriginalSQL          string
-	BindVariables        map[string]*querypb.BindVariable
-	rewrittenSqls        []string
+	BindVariables        map[string]streamlog.BindVariable `json:",omitempty"`
+	rewrittenSqls        []string                          `json:"-"`
 	RowsAffected         int
 	NumberOfQueries      int
 	StartTime            time.Time
@@ -57,11 +58,22 @@ type LogStats struct {
 	MysqlResponseTime    time.Duration
 	WaitingForConnection time.Duration
 	QuerySources         byte
-	Rows                 [][]sqltypes.Value
+	Rows                 [][]sqltypes.Value `json:"-"`
 	TransactionID        int64
 	ReservedID           int64
-	Error                error
+	Error                error `json:",omitempty"`
 	CachedPlan           bool
+}
+
+type LogStatsJSON struct {
+	CallInfo        string
+	Username        string
+	ImmediateCaller string
+	EffectiveCaller string
+	RewrittenSQL    string
+	TotalTime       time.Duration
+	ResponseSize    int
+	LogStats
 }
 
 // NewLogStats constructs a new LogStats with supplied Method and ctx
@@ -189,10 +201,12 @@ func (stats *LogStats) Logf(w io.Writer, params url.Values) error {
 
 		_, fullBindParams := params["full"]
 		formattedBindVars = sqltypes.FormatBindVariables(
-			stats.BindVariables,
+			streamlog.BindVariablesToProto(stats.BindVariables),
 			fullBindParams,
 			streamlog.GetQueryLogFormat() == streamlog.QueryLogFormatJSON,
 		)
+	} else {
+		stats.BindVariables = nil
 	}
 
 	// TODO: remove username here we fully enforce immediate caller id
@@ -204,6 +218,19 @@ func (stats *LogStats) Logf(w io.Writer, params url.Values) error {
 	case streamlog.QueryLogFormatText:
 		fmtString = "%v\t%v\t%v\t'%v'\t'%v'\t%v\t%v\t%.6f\t%v\t%q\t%v\t%v\t%q\t%v\t%.6f\t%.6f\t%v\t%v\t%v\t%q\t\n"
 	case streamlog.QueryLogFormatJSON:
+		if streamlog.UseQueryLogJSONV2() {
+			// flag --querylog-json-v2
+			return json.NewEncoder(w).Encode(LogStatsJSON{
+				EffectiveCaller: stats.EffectiveCaller(),
+				ImmediateCaller: stats.ImmediateCaller(),
+				LogStats:        *stats,
+				Username:        username,
+				CallInfo:        callInfo,
+				RewrittenSQL:    rewrittenSQL,
+				ResponseSize:    stats.SizeOfResponse(),
+				TotalTime:       stats.TotalTime(),
+			})
+		}
 		fmtString = "{\"Method\": %q, \"CallInfo\": %q, \"Username\": %q, \"ImmediateCaller\": %q, \"Effective Caller\": %q, \"Start\": \"%v\", \"End\": \"%v\", \"TotalTime\": %.6f, \"PlanType\": %q, \"OriginalSQL\": %q, \"BindVars\": %v, \"Queries\": %v, \"RewrittenSQL\": %q, \"QuerySources\": %q, \"MysqlTime\": %.6f, \"ConnWaitTime\": %.6f, \"RowsAffected\": %v,\"TransactionID\": %v,\"ResponseSize\": %v, \"Error\": %q}\n"
 	}