## Description

This PR adds a new version 2.0 format for `--querylog-format=json`. This new format is opt-in with the flag `--querylog-json-v2`, however I propose the new format replaces the current format in later releases

The motivation is to resolve cases where the [current format produces invalid JSON](https://github.com/vitessio/vitess/issues/12872) caused by certain bind variables values that break `json`'s spec _(bug confirmed/identified by Vinted and Slack in prod - [Slack :thread:](https://vitess.slack.com/archives/CMDJ2KFEZ/p1687883596023629))_

The version 2.0 format resolves the bind-variable breakage by letting `encoding/json` library marshal the `[]byte` value as base64, which is default behaviour in the library. The original implementation builds json manually using `fmt.Sprintf(...)`, printing `[]byte` as `%q`, which explains how we got here. The new format is a breaking change to users of `--querylog-format=json`, however, which is why I added this behind a flag

Breaking changes:
- Bind variable values are base64 encoded _(transparently by `encoding/json`)_
    - Resolves https://github.com/vitessio/vitess/issues/12872
    - Users reading querylogs with Go's `encoding/json` + unmarshal using the `LogStatsJSON` struct _(or the value field only as `[]byte`)_ won't need to know it was base64
    - Users reading querylogs from other languages will need to be aware of the base64 encoding of bind var values, but I think that is unavoidable
- `Effective Caller` field renamed to `EffectiveCaller`, to be consistent with `ImmediateCaller` and other camel-cased fields
    - Also spaces in json keys is kind of gross 😄 
- `Start` and `End` renamed to `StartTime` and `EndTime` respectively, for clarity and to be consistent with the other `Time`-suffixed vars _(eg: `PlanTime`/`ExecuteTime`)_
- Pre-existing `CachedPlan` key/value added to new `vttablet` format
- The order of the key/values changed
- When certain values are `nil`/`""` they are not added to the json _(via `json:",omitempty"`)_

The existing `text` and `json` _(without `--querylog-json-v2` flag)_ formats should be unaffected by this change

Example `vtgate` log _(pretty-printed, without redaction)_:
```json
{
  "RemoteAddr": "<ADDR HERE>",
  "Username": "<USERNAME HERE>",
  "ImmediateCaller": "<CALLER HERE>",
  "EffectiveCaller": "<CALLER HERE>",
  "Method": "test",
  "TabletType": "PRIMARY",
  "StmtType": "select",
  "SQL": "select * from testtable where name = :strVal and message = :bytesVal",
  "BindVariables": {
    "bytesVal": {
      "Type": "VARBINARY",
      "Value": "FmtAtEq6S9Y="
    },
    "strVal": {
      "Type": "VARCHAR",
      "Value": "YWJj"
    }
  },
  "StartTime": "2017-01-01T01:02:03Z",
  "EndTime": "2017-01-01T01:02:04.000001234Z",
  "ShardQueries": 0,
  "RowsAffected": 0,
  "RowsReturned": 0,
  "PlanTime": 0,
  "ExecuteTime": 0,
  "CommitTime": 0,
  "TablesUsed": [
    "ks1.tbl1",
    "ks2.tbl2"
  ],
  "SessionUUID": "suuid",
  "CachedPlan": false,
  "ActiveKeyspace": "db"
}
```

Example `vttablet` log _(pretty-printed, without redaction)_:
```json
{
  "CallInfo": "<INFO HERE>",
  "Username": "<USERNAME HERE>",
  "ImmediateCaller": "<CALLER HERE>",
  "EffectiveCaller": "<CALLER HERE>",
  "RewrittenSQL": "sql with pii",
  "TotalTime": 1000001234,
  "ResponseSize": 1,
  "Method": "test",
  "PlanType": "",
  "OriginalSQL": "sql",
  "BindVariables": {
    "bytesVal": {
      "Type": "VARBINARY",
      "Value": "FmtAtEq6S9Y="
    },
    "intVal": {
      "Type": "INT64",
      "Value": "MQ=="
    }
  },
  "RowsAffected": 0,
  "NumberOfQueries": 1,
  "StartTime": "2017-01-01T01:02:03Z",
  "EndTime": "2017-01-01T01:02:04.000001234Z",
  "MysqlResponseTime": 0,
  "WaitingForConnection": 0,
  "QuerySources": 2,
  "TransactionID": 12345,
  "ReservedID": 0,
  "CachedPlan": false
}
```

## Related Issue(s)

1. https://github.com/vitessio/vitess/issues/12872
2. https://github.com/vitessio/vitess/pull/12988

## Checklist

-   [x] "Backport to:" labels have been added if this change should be back-ported to release branches
-   [x] If this change is to be back-ported to previous releases, a justification is included in the PR description
-   [x] Tests were added or are not required
-   [x] Did the new or modified tests pass consistently locally and on CI?
-   [x] Documentation was added or is not required

## Deployment Notes

<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->