@vmg 2024-04-01
Oy my bad, I completely overlooked this PR when I began working on https://github.com/vitessio/vitess/pull/15539 so there's an obvious conflict. That PR is now merged and the `arewefast` results are looking _amazingly good_ (up to 5% less CPU usage on the tablets in the OLTP benchmarks, which are usually not very CPU hungry). So now we had a good understanding that query logging is a CPU bottleneck, and the approach I chose in #15539 seems to be very good at tackling that whilst being reasonably maintainable.

Because of this, I'd like to see the changes in this PR implemented using the same style of "log builder" that I've merged. I think all the proposed syntax changes here are perfectly reasonable, and shipping them behind a `v2` flag is a good idea, but I'd be very wary of using `encoding/json` as the serialization because it allocates like a sieve. It'd be a significant regression!

@timvaillancourt: I'm gonna try to find some time this week (maybe Friday morning since it's a day off) to port this PR to the new builder API. If this is urgent to you, please feel free to do it yourself and ask me any questions! Otherwise I don't mind picking up the work here because, again, we've found out how critical performance-wise is the logging path.