@timvaillancourt 2024-04-10
> Oy my bad, I completely overlooked this PR when I began working on #15539 so there's an obvious conflict. That PR is now merged and the `arewefast` results are looking _amazingly good_ (up to 5% less CPU usage on the tablets in the OLTP benchmarks, which are usually not very CPU hungry). So now we had a good understanding that query logging is a CPU bottleneck, and the approach I chose in #15539 seems to be very good at tackling that whilst being reasonably maintainable.

@vmg 👋 long time! No problem, I'm excited to see performance improvements - I suspected this code path could be optimized
 
> I'd be very wary of using `encoding/json` as the serialization because it allocates like a sieve. It'd be a significant regression!

Yeah, that's a big drawback for this PR 👎. Perhaps this is why `fmt.Sprintf(...)` was used originally? I wonder if this could eventually move to `github.com/go-json-experiment/json`, once it's part of the standard library

> @timvaillancourt: I'm gonna try to find some time this week (maybe Friday morning since it's a day off) to port this PR to the new builder API. If this is urgent to you, please feel free to do it yourself and ask me any questions! Otherwise I don't mind picking up the work here because, again, we've found out how critical performance-wise is the logging path.

This isn't urgent but happy to help somehow! If you have some time to describe what you had planned I can offload this optionally