@DeathBorn 2024-04-17
https://github.com/vitessio/vitess/pull/15539/files#diff-3c26e166a7ecbd772596c75f87d03b0dc849be5a5a443b8429be0e28a170519cR66-R68
@vmg I think there is a bug - it should be `Type` and `Value` instead of lowercased values.