@DeathBorn 2024-04-17
@timvaillancourt from our perspective, it is good enough to encode only JSON string values, so this is my backport on top of almost zero allocation logger.
https://github.com/vitessio/vitess/compare/a1f1266f1933ed5a5f40b28a9eecb8c6a39125dc...a774f11e4050bef13aadeb1e7ce5619cf4829f9f


<details>
  <summary><i>encode bench</i></summary>

```go

import (
	"bytes"
	"context"
	"fmt"
	"testing"
	"time"

	"vitess.io/vitess/go/logstats"
	"vitess.io/vitess/go/streamlog"
	querypb "vitess.io/vitess/go/vt/proto/query"
	"vitess.io/vitess/go/vt/vtgate"
)

func BenchmarkLogStatsFormat(b *testing.B) {
	b.ReportAllocs()

	stats := vtgate.NewLogStats(context.Background(), "test", "sql1", "suuid", nil)
	stats.StartTime = time.Date(2017, time.January, 1, 1, 2, 3, 0, time.UTC)
	stats.EndTime = time.Date(2017, time.January, 1, 1, 2, 4, 1234, time.UTC)

	stats.Table = "ks1.tbl1,ks2.tbl2"
	stats.TabletType = "MASTER"
	stats.Keyspace = "db"
	stats.BindVariables = map[string]*querypb.BindVariable{
		"Type_BINARY,  ": {Type: querypb.Type_BINARY, Value: []byte("a")},                     // encoded base64
		"Type_BIT,     ": {Type: querypb.Type_BIT, Value: []byte("1")},                        // encoded base64
		"Type_BLOB,    ": {Type: querypb.Type_BLOB, Value: []byte("a")},                       // encoded base64
		"Type_CHAR,    ": {Type: querypb.Type_CHAR, Value: []byte("a")},                       // encoded base64
		"Type_DATE,    ": {Type: querypb.Type_DATE, Value: []byte("2012-02-24")},              // encoded base64
		"Type_DATETIME,": {Type: querypb.Type_DATETIME, Value: []byte("2012-02-24 23:19:43")}, // encoded base64
		"Type_DECIMAL, ": {Type: querypb.Type_DECIMAL, Value: []byte("1.00")},                 // encoded base64
		"Type_ENUM,    ": {Type: querypb.Type_ENUM, Value: []byte("a")},                       // encoded base64
		"Type_FLOAT32, ": {Type: querypb.Type_FLOAT32, Value: []byte("1.00")},
		"Type_FLOAT64, ": {Type: querypb.Type_FLOAT64, Value: []byte("1.00")},
		"Type_INT8,    ": {Type: querypb.Type_INT8, Value: []byte("1")},
		"Type_INT16,   ": {Type: querypb.Type_INT16, Value: []byte("1")},
		"Type_INT24,   ": {Type: querypb.Type_INT24, Value: []byte("1")},
		"Type_INT32,   ": {Type: querypb.Type_INT32, Value: []byte("1")},
		"Type_INT64,   ": {Type: querypb.Type_INT64, Value: []byte("1")},
		"Type_SET,     ": {Type: querypb.Type_SET, Value: []byte("a")},                         // encoded base64
		"Type_TEXT,    ": {Type: querypb.Type_TEXT, Value: []byte("a")},                        // encoded base64
		"Type_TIME,    ": {Type: querypb.Type_TIME, Value: []byte("23:19:43")},                 // encoded base64
		"Type_TIMESTAMP": {Type: querypb.Type_TIMESTAMP, Value: []byte("2012-02-24 23:19:43")}, // encoded base64
		"Type_UINT8,   ": {Type: querypb.Type_UINT8, Value: []byte("1")},
		"Type_UINT16,  ": {Type: querypb.Type_UINT16, Value: []byte("1")},
		"Type_UINT24,  ": {Type: querypb.Type_UINT24, Value: []byte("1")},
		"Type_UINT32,  ": {Type: querypb.Type_UINT32, Value: []byte("1")},
		"Type_UINT64,  ": {Type: querypb.Type_UINT64, Value: []byte("1")},
		"Type_VARBINARY": {Type: querypb.Type_VARBINARY, Value: []byte("a")}, // encoded base64
		"Type_VARCHAR, ": {Type: querypb.Type_VARCHAR, Value: []byte("a")},   // encoded base64
		"Type_YEAR,    ": {Type: querypb.Type_YEAR, Value: []byte("1")},      // encoded base64
		"Type_TUPLE,1  ": {Type: querypb.Type_TUPLE, Values: []*querypb.Value{{Type: querypb.Type_FLOAT64, Value: []byte("1")}, {Type: querypb.Type_FLOAT64, Value: []byte("2")}}},
		"Type_TUPLE,2  ": {Type: querypb.Type_TUPLE, Values: []*querypb.Value{{Type: querypb.Type_INT64, Value: []byte("1")}, {Type: querypb.Type_INT64, Value: []byte("2")}}},
		"Type_TUPLE,3  ": {Type: querypb.Type_TUPLE, Values: []*querypb.Value{{Type: querypb.Type_UINT64, Value: []byte("1")}, {Type: querypb.Type_UINT64, Value: []byte("2")}}},
		"Type_TUPLE,4  ": {Type: querypb.Type_TUPLE, Values: []*querypb.Value{{Type: querypb.Type_VARBINARY, Value: []byte("aa")}, {Type: querypb.Type_VARBINARY, Value: []byte("bb")}}},
		"Type_TUPLE,5  ": {Type: querypb.Type_TUPLE, Values: []*querypb.Value{{Type: querypb.Type_VARBINARY, Value: []byte("aa")}, {Type: querypb.Type_INT64, Value: []byte("1")}}},
	}
	fullBindParams := true
	remoteAddr, username := stats.RemoteAddrUsername()
	*streamlog.QueryLogFormat = streamlog.QueryLogFormatJSON
	*streamlog.QueryLogJSONV2 = true

	b.RunParallel(func(pb *testing.PB) {
		for pb.Next() {
			var w bytes.Buffer
			log := logstats.NewLogger()
			log.Init(*streamlog.QueryLogFormat == streamlog.QueryLogFormatJSON)
			log.Key("Method")
			log.StringUnquoted(stats.Method)
			log.Key("RemoteAddr")
			log.StringUnquoted(remoteAddr)
			log.Key("Username")
			log.StringUnquoted(username)
			log.Key("ImmediateCaller")
			log.StringSingleQuoted(stats.ImmediateCaller())
			log.Key("Effective Caller")
			log.StringSingleQuoted(stats.EffectiveCaller())
			log.Key("Start")
			log.Time(stats.StartTime)
			log.Key("End")
			log.Time(stats.EndTime)
			log.Key("TotalTime")
			log.Duration(stats.TotalTime())
			log.Key("PlanTime")
			log.Duration(stats.PlanTime)
			log.Key("ExecuteTime")
			log.Duration(stats.ExecuteTime)
			log.Key("CommitTime")
			log.Duration(stats.CommitTime)
			log.Key("StmtType")
			log.StringUnquoted(stats.StmtType)
			log.Key("SQL")
			log.String(stats.SQL)
			log.Key("BindVars")
			if *streamlog.QueryLogJSONV2 {
				log.BindVariablesV2(stats.BindVariables, fullBindParams)
			} else {
				log.BindVariables(stats.BindVariables, fullBindParams)
			}
			log.Key("ShardQueries")
			log.Uint(stats.ShardQueries)
			log.Key("RowsAffected")
			log.Uint(stats.RowsAffected)
			log.Key("Error")
			log.String(stats.ErrorStr())
			log.Key("TabletType")
			log.String(stats.TabletType)
			log.Key("SessionUUID")
			log.String(stats.SessionUUID)
			log.Key("InTransaction")
			log.Bool(stats.InTransaction)
			log.Key("Table")
			log.String(stats.Table)
			log.Key("Keyspace")
			log.String(stats.Keyspace)

			log.Flush(&w)
		}
	})

}
```
</details>


<details>
  <summary><i>old bench</i></summary>

```go

import (
	"bytes"
	"context"
	"fmt"
	"testing"
	"time"

	"vitess.io/vitess/go/sqltypes"
	"vitess.io/vitess/go/streamlog"
	querypb "vitess.io/vitess/go/vt/proto/query"
	"vitess.io/vitess/go/vt/vtgate"
)

func BenchmarkLogStatsFormat(b *testing.B) {
	b.ReportAllocs()

	stats := vtgate.NewLogStats(context.Background(), "test", "sql1", "suuid", nil)
	stats.StartTime = time.Date(2017, time.January, 1, 1, 2, 3, 0, time.UTC)
	stats.EndTime = time.Date(2017, time.January, 1, 1, 2, 4, 1234, time.UTC)

	stats.Table = "ks1.tbl1,ks2.tbl2"
	stats.TabletType = "MASTER"
	stats.Keyspace = "db"
	stats.BindVariables = map[string]*querypb.BindVariable{
		"Type_BINARY,  ": {Type: querypb.Type_BINARY, Value: []byte("a")},
		"Type_BIT,     ": {Type: querypb.Type_BIT, Value: []byte("1")},
		"Type_BLOB,    ": {Type: querypb.Type_BLOB, Value: []byte("a")},
		"Type_CHAR,    ": {Type: querypb.Type_CHAR, Value: []byte("a")},
		"Type_DATE,    ": {Type: querypb.Type_DATE, Value: []byte("2012-02-24")},
		"Type_DATETIME,": {Type: querypb.Type_DATETIME, Value: []byte("2012-02-24 23:19:43")},
		"Type_DECIMAL, ": {Type: querypb.Type_DECIMAL, Value: []byte("1.00")},
		"Type_ENUM,    ": {Type: querypb.Type_ENUM, Value: []byte("a")},
		"Type_FLOAT32, ": {Type: querypb.Type_FLOAT32, Value: []byte("1.00")},
		"Type_FLOAT64, ": {Type: querypb.Type_FLOAT64, Value: []byte("1.00")},
		"Type_INT8,    ": {Type: querypb.Type_INT8, Value: []byte("1")},
		"Type_INT16,   ": {Type: querypb.Type_INT16, Value: []byte("1")},
		"Type_INT24,   ": {Type: querypb.Type_INT24, Value: []byte("1")},
		"Type_INT32,   ": {Type: querypb.Type_INT32, Value: []byte("1")},
		"Type_INT64,   ": {Type: querypb.Type_INT64, Value: []byte("1")},
		"Type_SET,     ": {Type: querypb.Type_SET, Value: []byte("a")},
		"Type_TEXT,    ": {Type: querypb.Type_TEXT, Value: []byte("a")},
		"Type_TIME,    ": {Type: querypb.Type_TIME, Value: []byte("23:19:43")},
		"Type_TIMESTAMP": {Type: querypb.Type_TIMESTAMP, Value: []byte("2012-02-24 23:19:43")},
		"Type_UINT8,   ": {Type: querypb.Type_UINT8, Value: []byte("1")},
		"Type_UINT16,  ": {Type: querypb.Type_UINT16, Value: []byte("1")},
		"Type_UINT24,  ": {Type: querypb.Type_UINT24, Value: []byte("1")},
		"Type_UINT32,  ": {Type: querypb.Type_UINT32, Value: []byte("1")},
		"Type_UINT64,  ": {Type: querypb.Type_UINT64, Value: []byte("1")},
		"Type_VARBINARY": {Type: querypb.Type_VARBINARY, Value: []byte("a")},
		"Type_VARCHAR, ": {Type: querypb.Type_VARCHAR, Value: []byte("a")},
		"Type_YEAR,    ": {Type: querypb.Type_YEAR, Value: []byte("1")},
		"Type_TUPLE,1  ": {Type: querypb.Type_TUPLE, Values: []*querypb.Value{{Type: querypb.Type_FLOAT64, Value: []byte("1")}, {Type: querypb.Type_FLOAT64, Value: []byte("2")}}},
		"Type_TUPLE,2  ": {Type: querypb.Type_TUPLE, Values: []*querypb.Value{{Type: querypb.Type_INT64, Value: []byte("1")}, {Type: querypb.Type_INT64, Value: []byte("2")}}},
		"Type_TUPLE,3  ": {Type: querypb.Type_TUPLE, Values: []*querypb.Value{{Type: querypb.Type_UINT64, Value: []byte("1")}, {Type: querypb.Type_UINT64, Value: []byte("2")}}},
		"Type_TUPLE,4  ": {Type: querypb.Type_TUPLE, Values: []*querypb.Value{{Type: querypb.Type_VARBINARY, Value: []byte("aa")}, {Type: querypb.Type_VARBINARY, Value: []byte("bb")}}},
		"Type_TUPLE,5  ": {Type: querypb.Type_TUPLE, Values: []*querypb.Value{{Type: querypb.Type_VARBINARY, Value: []byte("aa")}, {Type: querypb.Type_INT64, Value: []byte("1")}}},
	}
	fullBindParams := true
	remoteAddr, username := stats.RemoteAddrUsername()
	*streamlog.QueryLogFormat = streamlog.QueryLogFormatJSON

	b.RunParallel(func(pb *testing.PB) {
		for pb.Next() {
			var w bytes.Buffer
			formattedBindVars := sqltypes.FormatBindVariables(
				stats.BindVariables,
				fullBindParams,
				*streamlog.QueryLogFormat == streamlog.QueryLogFormatJSON,
			)

			var fmtString string
			switch *streamlog.QueryLogFormat {
			case streamlog.QueryLogFormatText:
				fmtString = "%v\t%v\t%v\t'%v'\t'%v'\t%v\t%v\t%.6f\t%.6f\t%.6f\t%.6f\t%v\t%q\t%v\t%v\t%v\t%q\t%q\t%q\t%q\t%t\t%q\t\n"
			case streamlog.QueryLogFormatJSON:
				fmtString = "{\"Method\": %q, \"RemoteAddr\": %q, \"Username\": %q, \"ImmediateCaller\": %q, \"Effective Caller\": %q, \"Start\": \"%v\", \"End\": \"%v\", \"TotalTime\": %.6f, \"PlanTime\": %v, \"ExecuteTime\": %v, \"CommitTime\": %v, \"StmtType\": %q, \"SQL\": %q, \"BindVars\": %v, \"ShardQueries\": %v, \"RowsAffected\": %v, \"Error\": %q, \"Keyspace\": %q, \"Table\": %q, \"TabletType\": %q, \"InTransaction\": %t, \"SessionUUID\": %q}\n"
			}

			fmt.Fprintf(
				&w,
				fmtString,
				stats.Method,
				remoteAddr,
				username,
				stats.ImmediateCaller(),
				stats.EffectiveCaller(),
				stats.StartTime.Format("2006-01-02 15:04:05.000000"),
				stats.EndTime.Format("2006-01-02 15:04:05.000000"),
				stats.TotalTime().Seconds(),
				stats.PlanTime.Seconds(),
				stats.ExecuteTime.Seconds(),
				stats.CommitTime.Seconds(),
				stats.StmtType,
				stats.SQL,
				formattedBindVars,
				stats.ShardQueries,
				stats.RowsAffected,
				stats.ErrorStr(),
				stats.Keyspace,
				stats.Table,
				stats.TabletType,
				stats.InTransaction,
				stats.SessionUUID,
			)
		}
	})

}
```
</details>

```
// new
BenchmarkLogStatsFormat-10    	  361864	      3034 ns/op	    2594 B/op	      19 allocs/op
// old
BenchmarkLogStatsFormat-10    	  180460	      6335 ns/op	    9026 B/op	     113 allocs/op
```
Basically, only JSON string values add allocations. Tuples do not.